% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deMacron.R
\name{deMacron}
\alias{deMacron}
\title{deMacron a string}
\usage{
deMacron(original)
}
\arguments{
\item{string}{string to de-macron}
}
\description{
No, \code{deMacron} doesn't remove French politicians. It replaces all examples of macron vowels
in a string with the 'anglicised' equivalents. This is useful for analysing
strings such as hashtags where users might mix them but we want kakī == kaki for example. Macron
vowels are common in Māori - see https://kupu.maori.nz/about/macrons-keyboard-setup
}
\seealso{
Other Utilities: \code{\link{getLinuxMemProfile}},
  \code{\link{gzipIt}}, \code{\link{tidyNum}}
}
\author{
Ben Anderson, \email{banderson@soton.ac.uk}

David Hood
}
\concept{Utilities}
