% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalise.R
\name{equivalise}
\alias{equivalise}
\title{Equivalise a vector of numbers}
\usage{
equivalise(input, method)
}
\arguments{
\item{input}{an input vector}

\item{method}{the method to use (mclements, modified oecd, square root)}
}
\description{
\code{\link[dkutils:equivalise]{dkutils::equivalise()}} takes a vector of numbers (or a single number) and equivalises
it according to the method specified. Equivalisation is a way to compare attributes such
as income, expenditure or consumption across households of different sizes and
compositions. It is based on the idea that such values do not linearly increase
as .houseodl size increases and increase at different rates for different household
compositions. Equivalising the values means they ca be compared across household types.
For a detailed discussion and comparison of methods (not all of which are implemented
here) see \url{https://link.springer.com/article/10.1007/s00181-020-01822-6}
}
\details{
Methods supported:
\itemize{
\item Modified OECD  - moecd
`* McClements - mcl
\item Square root - sqrt
}
}
\examples{
equivalise(var = 123456789, method = "moecd")
equivalise(var = income, method = "mcl")

}
\seealso{
Other Utilities: 
\code{\link{deMacron}()},
\code{\link{getLinuxMemProfile}()},
\code{\link{gzipIt}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{banderson@soton.ac.uk}
}
\concept{Utilities}
