% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinuxMemProfile.R
\name{getLinuxMemProfile}
\alias{getLinuxMemProfile}
\title{Report current hardware profile - only works on Linux (not Mac OS X etc)}
\usage{
getLinuxMemProfile()
}
\description{
Report current hardware profile - only works on Linux (not Mac OS X etc)
}
\seealso{
Other Utilities: 
\code{\link{deMacron}()},
\code{\link{equivalise}()},
\code{\link{gzipIt}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{banderson@soton.ac.uk}
}
\concept{Utilities}
